moved calibration to generated calibration.py
